Revert "[Change] temporarily changed all time capture to double"
This reverts commit 992fcc67d13667c6128cd29e51ab1fb8e6a1d854.

+++ Args/Args-Editor/Header/Systems/TestMonoUpdateSystem.h
@@ -7,32 +7,32 @@ using namespace std;
 class TestMonoUpdateSystem : public Args::MonoUpdateSystem<TestMonoUpdateSystem>
 {
 private:
-	double accumDeltaTime = 0;
+	float accumDeltaTime = 0;
 	int updatesSincePrint = 0;
-	double elapsedTime = 0;
+	float elapsedTime = 0;
 
-	double totalQuery = 0.0;
-	double totalAdd = 0.0;
-	double totalRemove = 0.0;
+	float totalQuery = 0.f;
+	float totalAdd = 0.f;
+	float totalRemove = 0.f;
 
-	double lowestQueryFrame = DBL_MAX, highestQueryFrame = 0.0;
-	double lowestAddFrame = DBL_MAX, highestAddFrame = 0.0;
-	double lowestRemoveFrame = DBL_MAX, highestRemoveFrame = 0.0;
+	float lowestQueryFrame = FLT_MAX, highestQueryFrame = 0.f;
+	float lowestAddFrame = FLT_MAX, highestAddFrame = 0.f;
+	float lowestRemoveFrame = FLT_MAX, highestRemoveFrame = 0.f;
 
 	Args::uint32 totalTortured = 0;
 	Args::uint32 totalFrames = 0;
 	
 public:
-	static double lowestQuery;
-	static double highestQuery;
-	static double lowestAdd;
-	static double highestAdd;
-	static double lowestRemove;
-	static double highestRemove;
+	static float lowestQuery;
+	static float highestQuery;
+	static float lowestAdd;
+	static float highestAdd;
+	static float lowestRemove;
+	static float highestRemove;
 
-	static double addTime;
-	static double removeTime;
-	static double queryTime;
+	static float addTime;
+	static float removeTime;
+	static float queryTime;
 	static int torturedComponents;
 
 	TestMonoUpdateSystem() : Args::MonoUpdateSystem<TestMonoUpdateSystem>() {}
@@ -41,9 +41,9 @@ public:
 
 	void Start();
 
-	void Update(double deltaTime);
-	void Print(double deltaTime);
-	void Shutdown(double deltaTime);
+	void Update(float deltaTime);
+	void Print(float deltaTime);
+	void Shutdown(float deltaTime);
 
 	void Exit(Args::ControllerID controller, Args::ActionState action);
 };
